Add chats section & favicon
I made the submit button go right next to the message input. In addition, I added a another column for chats that the user has (this would be our next challenge in making direct messages). In addition, as a little plus, I added an favicon so that our website at least has a little logo/icon hahahah